Update patch set 1
Patch Set 1:

Jozef -

The correction for post is good, but it would be good if we could collapse the number of endpoints that exist. For example we have these URLS which are the same, except for the tag on the right. What is not clear to me is what makes one different from another such that we need two different entries.

POST /config/opendaylight-inventory:nodes/node/{id}/          group
POST /config/opendaylight-inventory:nodes/node/{id}/          group-features

If we can combine these, or clarify their usage that would be good.

Patch-set: 1
Label: Code-Review=0