Update patch set 1
Patch Set 1: Code-Review-1

I don't see how this makes the API and more usable then it currently is. 

Please explain the need to introduce yet one more interface into the mix that doesn't do anything different from the base or derived interfaces.

Also - if we do go forward with this merge we need documentation in that class explaining the same thing that I asked above.

Patch-set: 1