Update patch set 1
Patch Set 1: Code-Review+1

Catching an swallowing the NumberFormatException is better than throwing it. I think this will allow my non-openflow code to succeed (as per the test case provided).

It would be nicer if the exception was avoided and not thrown or swallowed-and-logged. The approach taken is undesirable because it adds noise to the log and stops the debugger at any breakpoint for NFE.

If the node is genuinely for openflow but id not numeric then it is not clear to me if the outcome is correct. There should be a test case with a non-numeric id for an openflow node (maybe there is already).

Patch-set: 1
Label: Code-Review=+1