Update patch set 1
Patch Set 1:

It is useful, because without this I would have to do casts to get the same interface as I get with .newReadWriteTransaction.

Without this change, the code in BGPCEP will be uglyfied. We need transaction chaining to convey our requirement to see state which we pushed asynchronously. Without that knowledge concurrent DS is not happening.

So, can we expedite things, *please*?

Patch-set: 1