Update patch set 1
Patch Set 1:

Ok. Tony your explanation makes sense. My understand is that you want to write code that doesn't care wether you are using a TransactionChain or a DataBroker.

Given Roberts request to move this forward as it is blocking a project I will push it, HOWEVER a few comments:

1) The commit message on changes like this NEED to provide context for reviews. Just looking at the code without the context (as in I want to write code that can use these two services without caring which one I am using) is critical to understanding why a change was made. So the commit message in future changes should be enhanced to provide this information.

2) This interface NEEDS comments as it is a public interface. So we need to follow up this patch set with one that adds detailed comments to the interface.

Patch-set: 1