Update patch set 1
Patch Set 1:

Hi Rob Adams,

This patch focus on Bug 1338 which is a bug that lists of actions are not ordered consistently.

I think the problem of "instruction-list" should be fixed in other patch after a bug report is created about it.

So this patch looks good to me as is.

Patch-set: 1