DO NOT MERGE - start on wiring sal-rest-connector
OK... this works entirely *except* for getting
the opendaylight/distribution/pom.xml
to copy the file into the configuration/initial
directory.

Please see integration/distributions/base/pom.xml
for an example of how to do that (note: I'm tryign to
here, but have yet to debug why its not working).

Change-Id: If826bb7352101833acd288b81cbcf7cec2f28b43
Signed-off-by: Ed Warnicke <eaw@cisco.com>