Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

One minor tweak to documentation and this is good.

In general, it would be great if we could minimize the formatting changes in patches so that we don't have them show up in the changes. If you really want to update formatting, ideally that should be done as a separate patch which is *only* formatting.

Patch-set: 2
Label: Code-Review=+1