Update patch set 2
Patch Set 2: -Code-Review

Hi Ed,

Thank you for review and comments.

Please note that when you build a controller on controller.git, the artifact of the build does not have the MD-SAL OF plugin, but it has the AD-SAL OF plugin (the legacy OF plugin) and sal-compatibility.
Therefore, just simply prioritizing sal-compatibility rather than the AD-SAL plugin does not fix the problem, because when the MD-SAL OF plugin is not installed, the sal-compatibility cannot do anything.

When the MD-SAL OF plugin is not installed, a controller should prioritize the AD-SAL OF plugin rather than the sal-compatibility.
This is what this patch does.


Your 1st proposal is that when the AD-SAL OF plugin, the MD-SAL OF plugin, and the sal-compatibility all are installed in a controller, then the controller should prioritize the sal-compatibility to use the MD-SAL plugin rather than the AD-SAL OF plugin, right?


Does your 2nd proposal imply that the sal-compatibility should be filtered out when the MD-SAL OF plugin is not installed and the AD-SAL OF plugin is installed?

Patch-set: 2