Update patch set 6
Patch Set 6: Code-Review-1

The ADSAL NSFs have a bad inter-dependencies amongst each other and the current set of Feature splits will not work.
We can either consider all the interdependent (bidirectional dependencies) bundles as a whole and install them as odl-nsf-managers and not expose individual features (or) redesign the NSFs to become proper components.

I would prefer to go with the first option of going with odl-nsf-managers wrapping up these bidirectional dependent managers and redesign them later if there is a real need.

Patch-set: 6
Label: Code-Review=-1