Update patch set 3
Patch Set 3: Code-Review-1

My suggestion is to *not* throw and *not* swallow the NFE. It is better to avoid the exceptional situation.

As implemented, the code should work for my usage but it is not admirable. 

From the book Effective Java by Josh Bloch:

Exceptions are, as their name implies, to be used only
for exceptional conditions; they should never be used for ordinary control flow.

Patch-set: 3
Label: Code-Review=-1