Update patch set 6
Patch Set 6:

I added the hosttracker.shell.version in commons/opendaylight/pom.xml

The tests verify that that when a shell command is called, the right string is returned.

I added the shell into odl-nsf-managers so that it is installed automatically when this feature is installed.

Patch-set: 6
Reviewer: Gerrit User 1950 <1950@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0