Update patch set 4
Patch Set 4:

Hi Tony,

I would like to point out here that 
section 5.6.4.1 of RFC 6020 says:

"A server MUST advertise all revisions of all modules it implements."

But the grammar in section 5.6.4 shows it as optional and, further, the "revision" statement in the YANG module is also optional (section 7.1.1) for the very first version of a module.  Thereafter, per section 10 of RFC6020, any revision of the YANG module MUST have a "revision" statement.

For YANG 1.0, I believe that your interpretation is correct, but also that this is effectively a bug in the YANG spec.

and does the patch 9490, enable ODL to read the yang file without revision? if yes then it's fine otherwise we need to have my proposed patch in ODL.


let me know your view on this.

Patch-set: 4
Reviewer: Gerrit User 1965 <1965@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0