BUG 1440 - Implement JSON to normalized node writer
JsonParserStream contains implementation of read method in which there
is gradually constructed structure of NodeDataWithSchema nodes.

NodeDataWithSchema is similar to SchemaNode but contains only reference to
children which really exists in JSON input and concrete values for nodes
of type leaf or leaf list.

It was necessary to firstly load JSON to NodeDataWitchSchema structure
because for some types (composite keys of list, all nodes belonging to one
augment, all nodes belonging to one choice) it isn't possible to do
sequential processing (it means to call concrete opening and closing
methods of NormalizedNodeStreamWriter interface).

Change-Id: Ie6b2c6e54582b22c31f30ac153f06d11deadff5f
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>