Update patch set 5
Patch Set 5:

There is an outstanding question around usability. This change simplifies the models which makes it a little clearer what values to post, but it adds many duplicate rest lines to the swagger which decreases usability in that aspect.

After discussing with Jozef via e-mail I think we need to get some more input from people on which is more usable.

Patch-set: 5
Label: Verified=0