BUG-1493: expose the deepest scope required by listeners
Recursive resolution of changes requires us to understand how deep
resolution is required by listeners attached to a particular node.

While we could calculate it in the sole user of this API,
ResolveDataChangeEventsTask, it would end up recalculating over and over
a value which changes only when the registered listeners change.

Thus, we maintain this information in ListenerTree.Node and expose it
via a new method, getDeepestScope().

While we're here, we move out DataChangeListenerRegistrationImpl, since
it is not visible to users. We also add additional protections to the
listener list, by exposing only an immutable view of it.

Change-Id: I80a1e36bc280c7767a581562e85789b9ba28ad1e
Signed-off-by: Robert Varga <rovarga@cisco.com>