Update patch set 12
Patch Set 12:

Robert - can we address your comments re: config with follow-up patches? This patch has been sitting for a couple weeks and I'd really like to get it merged so we can run performance tests with it. Devin has reviewed it as well and is ready to merge it today (he didn't want to merge it on a Fri).

Patch-set: 12
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0