Update patch set 12
Patch Set 12: Code-Review+2

Discussed with Robert - we will push as is and move the system properties to config properties in a future gerrit.

Do NOT rely on the system properties to exist in the future.

Patch-set: 12
Reviewer: Gerrit User 1840 <1840@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2
Label: SUBM=+1