Moved the resources to a separate plugin so they can be used by others.
The idea here would be to allow the projects to build
their own local runnable instances, as well as integration,
while keeping things in sync.

A couple of notes:
1)  Please be patient, this is a pull apart of work to
get features working in general, so there may be
some extra things in commons/opendaylight/pom.xml
for that are for that follow on patch.  Mea culpa.

2)  I created an opendaylight-karaf-empty distribution
that *just* pulls in the resources to be used
for feature testing (see next patch)

Change-Id: Id4816f7083cedde8fe1e1ff2aaa58d581d45c8ab
Signed-off-by: Ed Warnicke <eaw@cisco.com>