Update patch set 3
Patch Set 3:

I feel like we should really have JUnits here. I understand this is blocking the ability for more people to get started with karaf, so I am willing to push it forward without unit tests, but I would like a bug to be opened to follow it up.

Patch-set: 3