Update patch set 6
Patch Set 6:

(2 comments)

Not a specific comment on this patch but there are zero unit tests for inventory manager. This makes it harder for someone who is less familiar to go in an modify things because there is no safety net. 

Since this code is being modified and presumably tested on the controller, this may be a good time to add a few unit tests as a safety net.

Patch-set: 6
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0