Update patch set 2
Patch Set 2:

The code looks good to me. So +1 for that. I want Robert or someone else more familiar with the config sub system to look at it. If no one else takes a look before Monday AM, then I will push this.

Tom - two questions / requests preferably with this patch - or possible in a follow-up:

1) Can you modify some test to pass a real InMemoryDOMDataStoreCOnfigProperties object with non-default values and make sure they take? Just to ensure we don't ever regress there.
2) It seems prudent to sanity check the values (perhaps in the InMemoryDomDataStoreCOnfigProperties) to make sure the values make sense. I know we don't respect default values in restconf - I am not sure if the config sub system respects them either.

Patch-set: 2