Update patch set 2
Patch Set 2:

I missed that Moiz had already given this a +1 before the rebase. 

Tom P will add the additional Unit tests and then I will merge.

As for default values, Tom P pointed out that the values are defined as units, so we don't need to worry about negatives. So we are all set there.

Patch-set: 2