Adding clustering test app
This app has a three models - cars, people and a map of car and person

We have a provider for people to handle add a person rpc, this make sure we register that person's instance identifier for routing purpose.
Another provider is for car purchase, which handles routed rpc, this provider will just generate notification with person id and car id.

The notification listener will handle the datastore interaction and will use above notification data to add entry.

Change-Id: I2426eae2d417f51f1fa1eb7338839386b1651e66
Signed-off-by: Harman Singh <harmasin@cisco.com>