Update patch set 9
Patch Set 9:

Jozef - quick code overview looks good.

When I installed this we had a /config in the opendaylight-inventory second, but NOT in the toaster. One could argue that adding the post to /config for toaster is an enhancement and we have solved the bug by removing the bogus post to /config/toaster:toaster, BUT if it is possible to fix the toaster quickly so that post to /config is visible then that would be cool.

Let me know if you can try the fix or if we need to create a new bug for that request and push this fix.

Patch-set: 9
Label: Verified=0