Bug 1598: Cleanup stale ShardReadTransactions
For read-only Tx, a FinalizablePhantomReference is created with the
TransacionProxy instance as the referent and is added to a static
FinalizableReferenceQueue. The FinalizablePhantomReference is subclassed
to hold the TransactionProxy's remoteTransactionPaths map. When the
TransactionProxy instance in GC'ed, the FinalizablePhantomReference is
notified and calls closeTransaction on each TransactionContext to clean it up.

Change-Id: Ie644612cb34e7219dc089b8add6d397a11bffdda
Signed-off-by: tpantelis <tpanteli@brocade.com>