Update patch set 4
Patch Set 4:

Hi Ed, I think I am with Madhu here. This seems to be a blocker for some projects like OVSDB and VTN to get their features in Karaf so unless there is an absolute sureness this is going to break everything, I recommend to merge it asap. 

Regarding the request to ask 27 projects to download the patch and test before merging, I have never seen MD-SAL or yangtools people doing the same when they change the models, the datastore or anything that breaks everybody, so lets be fair here too.

If this patch has the risk of impacting some people I would encourage to send a note like "we have change this .... if someone experience any issue please be aware and contact ...". I wish I had seen a similar note at least once when other people break the controller.

BR/Luis

Patch-set: 4