Update patch set 4
Patch Set 4:

Thanks Luis for voicing your opinion.
This is a patch update on Tomcat (Not Jersey or Jackson or any other functional bundles that might impact the runtime differently for different usecases). Since being a Tomcat
Patch update, and since we have run quite a few tests on Controller, Integration and OVSDB projects, the level of confidence is extremely high that it will not impact any others.

Infact it has a positive effect that we have removed the Cors hack we added in Hydrogen time-frame.

So, am much comfortable to have this merged ASAP so that other projects can make progress.

Patch-set: 4