Update patch set 4
Patch Set 4:

Though we could potentially have karaf and legacy have different versions, based on Ryan's input on vulnerability analysis, it is infact better to upgrade to 7.0.53 asap even for the legacy distribution.

Also, this is not an impacting fix for any of the project because of the fact that this is Tomcat and not the framewroks that run on top of tomcat.

Patch-set: 4
Label: Verified=0