Update patch set 5
Patch Set 5:

FWIW, I modified the patch based on Ed's comment on keeping the surface area of the changes to Karaf. And not impacting legacy distribution.

Patchset 5 takes care of that request.

With this, the legacy distribution is not impacted in anyway and the upgrade is done only for karaf distribution.

Patch-set: 5
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0