Create patch set 3
Patch-set: 3
Subject: Bug 1593 - Flow Statistics manager is updating store with incorrect key Statistics manager cache the metadata related to stats data of all the entities (flow,group, meter,queue) like when was their respective stats were updated in operational data store. If statistics manager see that stats are not being updated from last two stats fetch cycle, it consider it stale and remove it. Object that it uses as keys for hash map cache contains some variant data (counters/durations) that varies in every stats fetch cycle. It creats duplicate entry in the cache for the same entity. After two fetch cycle statistics manager find old entry stale and removes it, but in actual that's not the case. This exception is thrown when it tries to remove the stats related to next duplicate entry (which was already removed ). Fixed it by using invariant key for caching the metadata to avoid the duplication.
Commit: f64aa7a061075dffa4a38164183bdf90b4412f9b