Update patch set 1
Patch Set 1:

There's other synchronization in this class that should probably be reviewed.

I don't think onGlobalContextUpdated needs to be synchronized as the dataTree handles synchronization. This isn't a big deal since context updates are infrequent.

I'm curious why registerChangeListener synchronizes this. It doesn't need to protect the listenerTree as it has a locks internally. Also in the AbstractListenerRegistration anonymous class, it syncs on the enclosing InMemoryDOMDataStore instance when closing the reg - this looks a little odd.

Does ThreePhaseCommitImpl#commit need to sync this? An instance is created per Tx commit and commit is called by one thread and should never be called twice, let alone concurrently.  I do see a potential visibility issue with the 'candidate' and listenerResolver' instance fields. They're both set in preCommit() on an executor Callable and they're accessed in commit() which could be different threads. So it would be safest to make these volatile. Or try to piggyback their sync via another volatile 'guard'.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0