Update patch set 3
Patch Set 3:

(1 comment)

Cosmetic nit... start="true" is the default for bundles in a feature.
(I checked in with the karaf guys to be sure) and so it is not needed in any of the <bundles>.

Note: I know this was pre-existing.
Note: I don't consider this to be something to hold the patch over, but would be nice to get cleaned up lest folks copy it.

Patch-set: 3
Reviewer: Gerrit User 9 <9@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0