Update patch set 3
Patch Set 3:

Okay. I agree that this patch must be merged ASAP as there are various other features are dependent on it.
Also the Controller Jenkins jobs are running crazy now and it is draining the limited time we have.

So, please merge it. I will address it during times when jenkins is a bit more sane :)

Patch-set: 3
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0