Update patch set 3
Patch Set 3:

hmm.. All the comments were addressed. including the naming convention. I changed it from base-features to features-base as suggested by the naming convention.

Can you please let me know which 1 are you referring to ?

And. Calling this as adsal-base is not appropriate just because of the fact that these are NOT adsal specific (though some of them are used by ADSAL components).

Nothing stops from others to use them.
Infact The jackson is being used by southbound plugins such as OVSDB.

Patch-set: 3