Update patch set 1
Patch Set 1: Code-Review-1

Though it passes all the Karaf Guidelines tests, it doesnt pass the dependency / Business logic tests. Reasons :

1. Sal-Compatibility should not be part of adsal-all ... because it is specific to OPENFLOW (please refer to ComponentActivator).

2. By the nature of Sal-Compatibility & for its business logic to make any sense, it has dependency on the adsal-core (This dependency is missing in the feature definition)

3. sal-compatibility is awesome, but has a strong dependency on MD-SAL & just by installing adsal-all, now we will drag MD-SAL components along, which is NOT a mandatory part of adsal ecosystem.

4. Its known that sal-compatibility and AD-SAL openflow plugins dont play well with each other and we need to provide a way to keep these incompatible bundles isolated and hence would prefer sal-compatibility not part of odl-adsal-all.

So, in short, Thanks for taking care of sal-compatiblity. But I feel that it should not be part of odl-adsal-all. But an optional feature that can be enabled in the presence of MD-SAL and Openflowplugin.

Will that work for you ?

Patch-set: 1
Label: Code-Review=-1