Update patch set 6
Patch Set 6:

Good point Moiz. That crossed my mind too. The problem is with the SchemaContext. Whenever that changes we need to propagate the change throughout the system. The UpdateSchemaContext message serves that purpose however it wasn't complete - DataChangeListenerProxy instances weren't updated. There's pros and cons either way. Using the update message is probably more in line with akka methodology rather than using shared mutable state that can change out-of-band from akka messages. However using update messages is harder to maintain as more classes, actor or other otherwise, that need the SchemaContext are added - you have to keep them up-to-date. The shared ShardContext alleviates that. 

I mainly added the ShardContext to wrap config params and make it easier when adding new config params as we only have to add to the ShardContext and we don't have to change "props" method signatures down the chain. 

That said, I can look into removing SchemaContext from the ShardContext. The UpdateSchemaContext message is still in play as it's needed to update the Shard InMemoryDOMDataStore. So having 2 mechanisms to propagate SchemaContext change isn't ideal - I wasn't crazy about that. I added it to ShardContext for the DataChangeListenerProxy and for 1 less param passed to ShardTransaction#props. But the SchemaContext in DataChangeListenerProxy isn't actually used currently - we pass it to DataChange but we don't serialize DataChange instances nor pass them over the wire. It isn't a big deal to pass SchemaContext to ShardTransaction#props. If I do that and remove SchemaContext from ShardContext then ShardContext can be immutable.

Patch-set: 6
Label: Verified=0