Update patch set 9
Patch Set 9:

"I suspect the
difference is b/c the JDK compiler generates the anonymous class as
static since they're enclosed in static methods but the Eclipse compiler
doesn't."

That's correct. It seems to be a bug in JDK compiler. I opened a defect a few weeks ago. This is what specification says:

http://docs.oracle.com/javase/specs/jls/se8/html/jls-15.html#jls-15.9.5

===
An anonymous class declaration is automatically derived from a class instance creation expression by the Java compiler.

An anonymous class is never abstract (§8.1.1.1).

An anonymous class is always implicitly final (§8.1.1.2).

An anonymous class is always an inner class (§8.1.3); it is never static (§8.1.1, §8.5.1).
===

Patch-set: 9