Update patch set 2
Patch Set 2: Code-Review-1

This is not being used on the Controller legacy UI. But it was added as a sample for anyone to make use of to add their own Branding. (Hence the name Brand fragment).

So, I dont think it is appropriate to remove this bundle.

Patch-set: 2
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1