Update patch set 5
Patch Set 5:

The yang has defaults so it isn't necessary for them to be in the config/initial XML. If you need to change them you can do it via rest or netconf on the fly (painfully as you mentioned) or just add them to the config/initial file and restart. The latter is easier, especially if they're in there already. I can do that.

Patch-set: 5