Update patch set 12
Patch Set 12:

(12 comments)

Looks good overall - nice work on the common Config stuff. Also clever use of "become" to get the metering behavior - a clean way to do it. It looks like akka handles the circular reference and avoids memory leak if you don't call "unbecome".

I assume the distributed store will be converted to use the CommonConfig in a later patch. Hopefully this will obviate the need for DatastoreContext and having to pass it down.

I also assume the rest of the rpc config will be made available via the config system in a later patch.

Patch-set: 12
Label: Verified=0