Update patch set 1
Patch Set 1:

(2 comments)

Seems reasonable - will wait for Moiz's comments to get addressed.

There are some ToDo's, \r and copy rights that need to be cleaned up - it can come in a follow up patch but I would suggest doing it now.

Also - this doesn't break the cyclic dependency between the AAA project and the controller project (assuming the AAA project brings in controller code). So we still have a problem if that is the case. I have not had a chance to look at this in detail yet though.

Patch-set: 1
Label: Verified=0