Update patch set 3
Patch Set 3:

All looks good - waiting for final verify job to complete before merge.

Note: we should have a follow-up patch that comes and removes the TODOs in the pom.xml and features.xml. We don't have to block this patch on that though since this is a RC0 blocker.

Patch-set: 3