Update patch set 6
Patch Set 6:

(2 comments)

Basheer answers to your questions,

What happens to the transaction chain when a transaction is aborted on a transaction chain?

I believe you can continue submitting transactions but this is completely governed by IMDS

What happens when a transaction on a transaction chain results in optimized lock exception ?

I am not a 100% sure on the state of the transaction chain which is really something which IMDS handles but the consumer would need to take some action here which could include, closing the chain or retrying an operation or whatever

Where are the checks that no new transaction can be added when a transaction chain is already in ready state?

In IMDS. On the CDS we try to make sure that if we have a ready on a transaction outstanding we do not send another create transaction on a chain. This is done in TransactionChainProxy

Patch-set: 6
Label: Verified=0